Consistent naming for ACC with pedal
Somehow the term "PDL" was introduced as a substitution for "PCC" but it's neither self explanatory nor widely used.
I suggest removing it entirely and just use "PCC" everywhere. I think it's more intuitive term and as a newcomer myself I can say it's less likely to create confusion for anyone joining the project.
I'll replace it also on Tinkla in case this gets merged.